Updated .version, _help.lua, entity.lua, renderer.lua, sound.lua. Added Icons.
Fixed some setting descriptions and the owndoc section of the manual.

Properly tagged DamageSource as a string type.

Cleared up how renderer.isFirstPerson works for other clients.

Completely changed sound.playSound to not use the deprecated version of the function. The deprecated version is still included as an alternate definition.

Prepared the project for description images.